Allows optional saving of filter means and moves ParticleFilter to separate module
