+++ pyfilter/filters/__init__.py
@@ -1,3 +1,5 @@
 from .apf import APF
 from .sisr import SISR
 from .ukf import UKF
+from .base import BaseFilter, BaseKalmanFilter
+from .pf import ParticleFilter
