+++ pyfilter/filters/ukf.py
@@ -2,18 +2,20 @@ from .base import BaseKalmanFilter
 from ..uft import UnscentedFilterTransform, UFTCorrectionResult
 from ..utils import choose
 import torch
+from typing import Dict
 
 
 class UKF(BaseKalmanFilter):
-    def __init__(self, model, **kwargs):
+    def __init__(self, model, utfkwargs: Dict[str, object] = None, **kwargs):
         """
         Implements the Unscented Kalman Filter by van der Merwe.
-        :param kwargs: Any kwargs passed to UnscentedTransform
+        :param kwargs: Any kwargs passed to `BaseFilter`
+        :param utfkwargs: Any kwargs passed to `UnscentedFilterTransform`
         """
 
-        super().__init__(model)
+        super().__init__(model, **kwargs)
 
-        self._ut = UnscentedFilterTransform(model, **kwargs)
+        self._ut = UnscentedFilterTransform(model, **(utfkwargs or dict()))
         self._ut_res = None
 
     def initialize(self):
