+++ pyfilter/inference/kernels/mh.py
@@ -51,7 +51,7 @@ class ParticleMetropolisHastings(BaseKernel):
         """
 
         t_filt.reset().initialize().longfilter(self._y, bar=False)
-        return t_filt.result.loglikelihood.sum(dim=0) - filter_.result.loglikelihood.sum(dim=0)
+        return t_filt.result.loglikelihood - filter_.result.loglikelihood
 
     def _before_resampling(self, filter_: BaseFilter, stacked: torch.Tensor):
         """
@@ -100,7 +100,7 @@ class ParticleMetropolisHastings(BaseKernel):
             else:
                 filter_.ssm.exchange(toaccept, t_filt.ssm)
 
-            weights = normalize(filter_.result.loglikelihood.sum(dim=0))
+            weights = normalize(filter_.result.loglikelihood)
 
         return self
