+++ pyfilter/inference/smc2.py
@@ -1,7 +1,7 @@
 from .base import SequentialParticleAlgorithm
 from .kernels import ParticleMetropolisHastings, SymmetricMH, OnlineKernel
 from ..utils import get_ess
-from ..filters.base import ParticleFilter
+from ..filters import ParticleFilter
 from ..kde import KernelDensityEstimate, NonShrinkingKernel
 from ..module import TensorContainer
 from torch import isfinite, zeros_like
@@ -36,8 +36,8 @@ class SMC2(SequentialParticleAlgorithm):
         self._y.append(y)
 
         # ===== Perform a filtering move ===== #
-        self.filter.filter(y)
-        self._w_rec += self.filter.result._loglikelihood[-1]
+        _, ll = self.filter.filter(y)
+        self._w_rec += ll
 
         # ===== Calculate efficient number of samples ===== #
         ess = get_ess(self._w_rec)
@@ -83,7 +83,7 @@ class SMC2(SequentialParticleAlgorithm):
             raise Exception(f'Configuration only allows {self._max_increases}!')
 
         # ===== Create new filter with double the state particles ===== #
-        oldlogl = self.filter.result.loglikelihood.sum(dim=0)
+        oldlogl = self.filter.result.loglikelihood
         oldparts = self.filter.particles[-1]
 
         self.filter.reset()
@@ -96,7 +96,7 @@ class SMC2(SequentialParticleAlgorithm):
         self.filter.set_nparallel(self._w_rec.shape[0]).initialize().longfilter(self._y.tensors, bar=False)
 
         # ===== Calculate new weights and replace filter ===== #
-        self._w_rec = self.filter.result.loglikelihood.sum(dim=0) - oldlogl
+        self._w_rec = self.filter.result.loglikelihood - oldlogl
         self._increases += 1
 
         return self
@@ -129,9 +129,17 @@ class SMC2FW(SequentialParticleAlgorithm):
         self._w_rec = zeros_like(self._smc2._w_rec)
         return self
 
+    def _make_switch(self):
+        self._smc2.rejuvenate()
+        self._switched = True
+        self._logged_ess = self._smc2._logged_ess
+        self._iterator.set_description(str(self))
+
+        return self
+
     def _update(self, y):
         # ===== Whether to use SMC2 or new ===== #
-        if len(self._smc2._y) < self._switch:
+        if not self._switched and len(self._smc2._y) < self._switch:
             if self._smc2._iterator is None:
                 self._smc2._iterator = self._iterator
 
@@ -139,10 +147,7 @@ class SMC2FW(SequentialParticleAlgorithm):
 
         # ===== Perform switch ===== #
         if not self._switched:
-            self._smc2.rejuvenate()
-            self._switched = True
-            self._logged_ess = self._smc2._logged_ess
-            self._iterator.set_description(str(self))
+            self._make_switch()
 
         # ===== Check if to propagate ===== #
         if self._num_iters >= self._bl or (~isfinite(self._w_rec)).any():
@@ -151,8 +156,8 @@ class SMC2FW(SequentialParticleAlgorithm):
             self._w_rec[:] = 0.
 
         # ===== Perform a filtering move ===== #
-        self.filter.filter(y)
-        self._w_rec += self.filter.result._loglikelihood[-1]
+        _, ll = self.filter.filter(y)
+        self._w_rec += ll
         self._num_iters += 1
 
         # ===== Calculate efficient number of samples ===== #
